django-webapp/requirements.txt
# Django and core dependencies
Django>=5.1.3
djangorestframework>=3.14.0

# WebSocket support (Channels)
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# HTTP Server
gunicorn>=21.2.0

# Configuration and utilities
python-decouple>=3.8
whitenoise>=6.6.0

# Push notifications
pywebpush>=1.14.0

# Optional: MQTT (if needed)
# paho-mqtt>=1.6.1

# Additional utilities
Pillow>=10.0.0  # For image processing if needed
django-redis>=5.4.0  # For Redis caching

# Firebase Vision API
google-cloud-vision>=3.4.4
google-auth>=2.23.0

# Email support
django-ses>=3.0.0  # For AWS SES (or use django.core.mail for SMTP)
sendgrid-django>=5.3.0  # For SendGrid email service

# SMS support
twilio>=8.0.0  # For SMS notifications via Twilio

# Image processing for thumbnails
Pillow>=10.0.0  # Already included, but ensuring it's available

# Payment processing
stripe>=7.0.0  # For subscription billing

